backend/requirements.txt
# Flask and Extensions
Flask==3.0.0
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2

# Geolocation
geopy==2.4.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Utilities
requests==2.31.0